Adds test to triger #18
This issue happens when there is captured stdout (so nose gives us the
exception value as a string instead of an exception object) and we try to
instantiate it using the exception class (we do that as a workaround for #16).
If the exception needs more than one argument to be instantiated, it will
receive only one (the string with the captured output).

I still need to find a fix for this, but it may very well be a change in the
way we use traceback.print_exception since we don't seem to have access to the
original exception arguments here.